fix: float initial time

+++ include/clock_animations.h
@@ -9,6 +9,6 @@ extern Animation seconds_hand;
 extern Animation minutes_hand;
 extern Animation hours_hand;
 
-void animation_start_at(Animation *anim, uint64_t now_ms, uint8_t start_value, uint8_t max_value);
+void animation_start_at(Animation *anim, uint64_t now_ms, float start_value, float max_value);
 
 #endif // CLOCK_ANIMATIONS_H
